Deleted endpoint and test added

+++ src/main/java/com/torresj/community/services/impl/UserServiceImpl.java
@@ -120,7 +120,9 @@ public class UserServiceImpl implements UserService, UserDetailsService {
 
     @Override
     public void delete(long userId) {
-        repository.deleteById(userId);
+        if (repository.existsById(userId)) {
+            repository.deleteById(userId);
+        }
     }
 
     @Override
